examples/blast_radius/pyproject.toml
[project]
name = "blast-radius"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.43",
    "httpx>=0.27.2",
    "matplotlib>=3.9.2",
    "networkx>=3.4.2",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
    "ruff>=0.7.0",
    "scipy>=1.14.1",
    "tenacity>=9.0.0",
    "unidiff>=0.7.5",
]
